requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

langchain==0.2.1
langchain-groq==0.1.6
langchain-community==0.2.1
langchain-core==0.2.3
langsmith==0.1.77

sentence-transformers==2.7.0

torch==2.3.0
transformers==4.41.1

chromadb==0.5.0

ragas==0.1.9
datasets==2.19.1

loguru==0.7.2

pypdf==4.2.0
tiktoken==0.7.0

pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
